backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
networkx
neo4j
python-dotenv
pytest
